requirements.txt
opencv-python>=4.8
mediapipe>=0.10
numpy>=1.24
pyaudio>=0.2.13
librosa>=0.10
SpeechRecognition>=3.10
Pillow>=10.0
matplotlib>=3.7
webrtcvad>=2.0.10
